third_party/blink/web_tests/virtual/text-antialias/custom-font-data-crash2-expected.txt

CONSOLE WARNING: Failed to decode downloaded font: custom-font-data-crash2.html
CONSOLE WARNING: OTS parsing error: invalid version tag
CONSOLE WARNING: line 15: Failed to decode downloaded font: custom-font-data-crash2.html
CONSOLE WARNING: line 15: OTS parsing error: invalid version tag
PASS
